tweaks and renames
