+++ src/main.tsx
@@ -8,7 +8,7 @@ import {
 } from 'react-router-dom';
 import { App } from './App';
 import './index.css';
-import './components/sidebar/hej.scss'
+import './components/burger-menu/hej.scss'
 
 const container = document.getElementById('root');
 const root = createRoot(container!);
